feat: enhance alpha evaluation and health calculation services with checklist handling and project health endpoints

+++ src/EssenceMvp.Mvc/Models/AlphaStateResult.cs
@@ -2,6 +2,6 @@
 
 public class AlphaStateResult
 {
-    public int CurrentStateNumber { get; set; }
+    public short CurrentStateNumber { get; set; }
     public string CurrentStateName { get; set; } = "";
 }
